agent/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.6
jinja2==3.1.5
